+++ src/components/SlideShow/slideShow.css
